Keep journey tests and guidance outside tooling core scope
